Use a much simpler approach to upgrading the web app
