+++ packages/playground/remote/src/lib/offline-mode-cache.ts
@@ -40,6 +40,36 @@ export async function cachedFetch(request: Request): Promise<Response> {
 	return response;
 }
 
+export async function networkFirstFetch(request: Request): Promise<Response> {
+	const offlineModeCache = await promisedOfflineModeCache;
+	const cachedResponse = await offlineModeCache.match(request, {
+		ignoreSearch: true,
+	});
+
+	let response: Response | undefined = undefined;
+	try {
+		response = await fetch(request, {
+			cache: 'no-cache',
+		});
+	} catch (e) {
+		if (cachedResponse) {
+			return cachedResponse;
+		}
+		throw e;
+	}
+
+	if (response.ok) {
+		await offlineModeCache.put(request, response.clone());
+		return response;
+	}
+
+	if (cachedResponse) {
+		return cachedResponse;
+	}
+
+	return response;
+}
+
 /**
  * For offline mode to work we need to cache all required assets.
  *
@@ -80,8 +110,6 @@ export async function cacheOfflineModeAssetsForCurrentRelease(): Promise<any> {
  * `buildVersionPlugin` for more details on how it's generated.
  */
 export async function purgeEverythingFromPreviousRelease() {
-	// @TODO: Ensure an older service worker won't ever remove the assets of a newer service worker,
-	//        even if this is accidentally called in the older worker.
 	const keys = await caches.keys();
 	const oldKeys = keys.filter(
 		(key) => key.startsWith(CACHE_NAME_PREFIX) && key !== LATEST_CACHE_NAME
